requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
-e git+https://github.com/edx/acid-block.git@98aecba94ecbfa934e2d00262741c0ea9f557fc9#egg=acid-xblock  # via -r requirements/test.in
appdirs==1.4.4            # via fs, virtualenv
arrow==1.1.0              # via jinja2-time
attrs==20.3.0             # via pytest
binaryornot==0.4.4        # via cookiecutter
bok_choy==0.7.1           # via -r requirements/test.in
boto3==1.17.60            # via fs-s3fs
boto==2.49.0              # via -r requirements/base.in
botocore==1.20.60         # via boto3, s3transfer
certifi==2020.12.5        # via requests
chardet==4.0.0            # via binaryornot, requests
click==7.1.2              # via cookiecutter
cookiecutter==1.7.2       # via -r requirements/base.in
coverage==5.5             # via -r requirements/test.in, pytest-cov
ddt==1.4.2                # via -r requirements/test.in
distlib==0.3.1            # via virtualenv
django-pyfs==3.0          # via -r requirements/base.in
filelock==3.0.12          # via tox, virtualenv
fs-s3fs==1.1.1            # via -r requirements/base.in, django-pyfs
fs==2.4.13                # via django-pyfs, fs-s3fs, xblock
idna==2.10                # via requests
iniconfig==1.1.1          # via pytest
jinja2-time==0.2.0        # via cookiecutter
jinja2==2.11.3            # via cookiecutter, jinja2-time
jmespath==0.10.0          # via boto3, botocore
lazy==1.4                 # via -r requirements/base.in, acid-xblock, bok-choy
lxml==4.6.3               # via -r requirements/base.in, xblock
mako==1.1.4               # via acid-xblock
markupsafe==1.1.1         # via cookiecutter, jinja2, mako, xblock
mock==4.0.3               # via -r requirements/test.in
needle==0.5.0             # via bok-choy
nose==1.3.7               # via needle
packaging==20.9           # via pytest, tox
pillow==8.2.0             # via needle
pluggy==0.13.1            # via pytest, tox
poyo==0.5.0               # via cookiecutter
py==1.10.0                # via pytest, tox
pyparsing==2.4.7          # via packaging
pypng==0.0.20             # via -r requirements/base.in
pytest-cov==2.11.1        # via -r requirements/test.in
pytest-django==4.2.0      # via -r requirements/test.in
pytest-rerunfailures==9.1.1  # via -r requirements/test.in
pytest==6.2.3             # via pytest-cov, pytest-django, pytest-rerunfailures
python-dateutil==2.8.1    # via arrow, botocore, xblock
python-slugify==4.0.1     # via cookiecutter
pytz==2021.1              # via django, fs, xblock
pyyaml==5.4.1             # via xblock
requests==2.25.1          # via -r requirements/base.in, cookiecutter
s3transfer==0.4.2         # via boto3
selenium==3.4.1           # via -r requirements/test.in, bok-choy, needle
simplejson==3.17.2        # via -r requirements/base.in
six==1.15.0               # via bok-choy, cookiecutter, fs, fs-s3fs, python-dateutil, tox, virtualenv, xblock
sqlparse==0.4.1           # via django
text-unidecode==1.3       # via python-slugify
toml==0.10.2              # via pytest, tox
tox-battery==0.6.1        # via -r requirements/test.in
tox==3.23.0               # via -r requirements/test.in, tox-battery
urllib3==1.26.4           # via botocore, requests
virtualenv==20.4.4        # via tox
web-fragments==1.0.0      # via -r requirements/base.in, xblock
webob==1.8.7              # via -r requirements/base.in, xblock
xblock==1.4.0             # via -r requirements/base.in, acid-xblock

# The following packages are considered to be unsafe in a requirements file:
# setuptools
